# Conflicts:
#	Delicacy/Delicacy.csproj

+++ Manager/FoodSortManager.cs
@@ -49,7 +49,8 @@ namespace Manager
         public OutputModel Delete(string id)
         {
             int i;
-            CheckParameter.PageCheck(id, out i);
+            if (!int.TryParse(id, out i))
+                return OutputHelper.GetOutputResponse(ResultCode.ErrorParameter);
             if (Service.Delete(i))
             {
                 return OutputHelper.GetOutputResponse(ResultCode.OK);
